backend/requirements.txt
python-dotenv==0.21.1
django==3.2
djangorestframework==3.14.0
psycopg2==2.9.5
Pillow==9.5.0
django-filter==23.1
